adds changeSelector method
Allows the user to modify the selector before it’s used in the
publication.

This can be helpful for narrowing results to a certain column based on
the search string used or for encrypting the search string before it’s
used in the query.

I’m not sure how useful this is for others but I thought I would submit
the PR just in case you wanted to add it.